refactor: extract mls init function to module
